client/streamlit-ui/requirements.txt
streamlit
requests
Pillow
numpy
reportlab
pydantic
matplotlib
plotly
statsmodels
scipy
protobuf<5

# LLM + search
google-genai
google-generativeai==0.8.3
agno         
duckduckgo-search
ddgs

tenacity==9.0.0
opencv-python-headless
